Map Commanded's `prefix` event store config to schema
